requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.2
asn1crypto==1.5.1
azure-core==1.35.1
azure-storage-blob==12.26.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
cssselect2==0.8.0
Django==5.2.6
django-meta==2.3.0
django-mssql-backend==2.8.1
django-storages==1.14.6
fonttools==4.60.1
freetype-py==2.5.1
html5lib==1.1
idna==3.10
isodate==0.7.2
lxml==6.0.2
mssql-django==1.6
oscrypto==1.3.0
pillow==11.3.0
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyodbc==5.2.0
pypdf==6.1.3
pyphen==0.17.2
python-bidi==0.6.7
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.8.2
transbank-sdk==4.0.0
mercadopago==2.3.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
requests>=2.31.0